requirements.txt
beautifulsoup4
requests
urllib
python-dotenv
w3lib
validators